+++ platform/timer.h
@@ -5,6 +5,6 @@
 
 void* timer_thread(void *_arg);
 void timer_control_write(u8 val);
-void timer_val_write(int val);
+void timer_val_write(u8 val);
 u8 timer_val_read();
 void timer_init();
